pyproject.toml
[tool.poetry]
name = "ai-army"
version = "0.1.3"
description = "AI-Army: Multi-agent orchestration with CrewAI for GitHub-driven development workflow"
readme = "README.md"
authors = ["AI-Army Contributors"]
license = "MIT"
packages = [
  { include = "ai_army", from = "src" },
  { include = "scripts" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
anthropic = ">=0.40.0"
crewai = { extras = ["tools"], version = ">=0.80.0" }
langchain = ">=0.3.0"
langchain-anthropic = ">=0.2.0"
PyGithub = ">=2.1.0"
apscheduler = ">=3.10.0"
pydantic-settings = ">=2.0.0"
python-dotenv = ">=1.0.0"
pyyaml = ">=6.0"
paramiko = ">=3.0.0"
requests = ">=2.28.0"
click = ">=8.0.0"
sentence-transformers = ">=2.2.0"
chromadb = ">=1.1.0,<1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.0.0"

[tool.poetry.scripts]
ai-army = "ai_army.main:main"
deploy = "scripts.deploy.release_to_droplet:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
